+++ common/GameConfig.cpp
@@ -140,6 +140,7 @@ bool GameConfig::visit_vars(T visitor, bool is_save)
     result &= visitor(dash_faction_key, "Window Mode", wnd_mode);
     result &= visitor(dash_faction_key, "Disable LOD Models", disable_lod_models);
     result &= visitor(dash_faction_key, "Anisotropic Filtering", anisotropic_filtering);
+    result &= visitor(dash_faction_key, "Nearest Filtering", nearest_filtering);
     result &= visitor(dash_faction_key, "MSAA", msaa);
     result &= visitor(dash_faction_key, "FPS Counter", fps_counter);
     result &= visitor(dash_faction_key, "Max FPS", max_fps);