+++ CHANGELOG.md
@@ -52,6 +52,7 @@ Version 1.5.0
 - display joining player IP address in dedicated server console
 - add dedicated server option for forcing player character
 - add '-exe-path' launcher command line argument that allows running multiple dedicated servers using separate RF directories
+- add support for nearest neighbor texture filtering ('nearest_filtering' command)
 - fix sending crash reports
 - fix non-working face scrolling before first geomod use in custom levels (RF bug)
 - fix memory and VRAM leak when quick saving (RF bug)